+++ findex/src/main/java/com/example/findex/api/IndexDataApi.java
@@ -4,6 +4,7 @@ import com.example.findex.dto.indexdata.data.IndexDataDto;
 import com.example.findex.dto.indexdata.request.IndexDataCreateRequest;
 import com.example.findex.dto.indexdata.response.CursorPageResponseIndexDataDto;
 import com.example.findex.dto.indexdata.response.ErrorResponse;
+import com.example.findex.dto.indexdata.response.IndexChartDto;
 import com.example.findex.dto.indexdata.response.IndexPerformanceDto;
 import com.example.findex.dto.indexdata.response.RankedIndexPerformanceDto;
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,6 +19,7 @@ import java.util.List;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Tag(name = "지수 데이터 API")
@@ -112,7 +114,6 @@ public interface IndexDataApi {
   @ApiResponses({
       @ApiResponse(responseCode = "200", description = "관심 지수 성과 조회 성공", content = @Content(schema = @Schema(implementation = IndexPerformanceDto.class))),
       @ApiResponse(responseCode = "400", description = "잘못된 요청")})
-  @GetMapping("/performance/favorite")
   ResponseEntity<List<IndexPerformanceDto>> getIndexFavoritePerformanceRank(
       @Parameter(description = "기간 타입 (DAILY, WEEKLY, MONTHLY)", required = true, example = "DAILY") @RequestParam String periodType
   );
@@ -121,7 +122,6 @@ public interface IndexDataApi {
   @ApiResponses({
       @ApiResponse(responseCode = "200", description = "지수 성과 랭킹 조회 성공", content = @Content(schema = @Schema(implementation = RankedIndexPerformanceDto.class))),
       @ApiResponse(responseCode = "400", description = "잘못된 요청")})
-  @GetMapping("/performance/rank")
   ResponseEntity<List<RankedIndexPerformanceDto>> getIndexPerformanceRank(
       @Parameter(description = "기간 타입 (DAILY, WEEKLY, MONTHLY)", required = true, example = "DAILY") @RequestParam String periodType,
 
@@ -134,7 +134,6 @@ public interface IndexDataApi {
   @ApiResponses({
       @ApiResponse(responseCode = "200", description = "CSV 파일 다운로드 성공", content = @Content(mediaType = "text/csv")),
       @ApiResponse(responseCode = "400", description = "잘못된 요청")})
-  @GetMapping("/export/csv")
   ResponseEntity<InputStreamResource> getIndexDataCsv(
       @Parameter(description = "지수 정보 ID", example = "1") @RequestParam(value = "indexInfoId", required = false) Long indexInfoId,
 
@@ -146,4 +145,20 @@ public interface IndexDataApi {
 
       @Parameter(description = "정렬 방향 (asc 또는 desc)", example = "desc", required = false) @RequestParam(value = "sortDirection", required = false, defaultValue = "desc") String sortDirection
   );
+
+    @Operation(
+            summary = "지수 차트 데이터 조회",
+            description = "주어진 기간(periodType)과 지수 ID(indexInfoId)에 대한 차트 데이터를 조회합니다."
+    )
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "성공적으로 차트 데이터를 반환함",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = IndexChartDto.class))),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청 (예: 유효하지 않은 periodType)"),
+            @ApiResponse(responseCode = "404", description = "해당 indexInfoId에 대한 데이터 없음")
+    })
+    ResponseEntity<IndexChartDto> getIndexChart(
+            @PathVariable @Parameter(description = "지수 정보 ID", example = "1") int indexInfoId,
+            @RequestParam @Parameter(description = "조회 기간 타입 (예: daily, weekly, monthly)", example = "daily") String periodType
+    );
 }
