+++ findex/src/main/java/com/example/findex/controller/IndexDataController.java
@@ -9,6 +9,7 @@ import com.example.findex.dto.indexdata.response.IndexPerformanceDto;
 import com.example.findex.dto.indexdata.response.RankedIndexPerformanceDto;
 import com.example.findex.entity.IndexData;
 import com.example.findex.service.IndexDataService;
+import com.example.findex.service.IndexInfoService;
 import java.io.ByteArrayInputStream;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
@@ -34,8 +35,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class IndexDataController implements IndexDataApi {
 
   private final IndexDataService indexDataService;
+  private final IndexInfoService indexInfoService;
 
   @Override
+  @GetMapping("/performance/favorite")
   public ResponseEntity<List<IndexPerformanceDto>> getIndexFavoritePerformanceRank(
       @RequestParam String periodType
   ){
@@ -53,6 +56,7 @@ public class IndexDataController implements IndexDataApi {
     return ResponseEntity.status(HttpStatus.OK).body(dto);
   }
 
+  @Override
   @GetMapping("/{indexInfoId}/chart")
   public ResponseEntity<IndexChartDto> getIndexChart(
       @PathVariable int indexInfoId,
@@ -88,6 +92,7 @@ public class IndexDataController implements IndexDataApi {
   }
 
   @Override
+  @GetMapping("/export/csv")
   public ResponseEntity<InputStreamResource> getIndexDataCsv(
       @RequestParam(value = "indexInfoId",required = false) Long indexInfoId,
       @RequestParam(value = "startDate",required = false) LocalDate startDate,
@@ -95,9 +100,8 @@ public class IndexDataController implements IndexDataApi {
       @RequestParam(value = "sortField",required = false, defaultValue = "baseDate") String sortField,
       @RequestParam(value = "sortDirection",required = false, defaultValue = "desc") String sortDirection
   ) {
-    List<IndexData> indexDataList = indexDataService.findByFilters(indexInfoId, startDate, endDate, sortField, sortDirection);
 
-    String csvData = convertToCsv(indexDataList);
+    String csvData = indexDataService.findToCsv(indexInfoId, startDate, endDate, sortField, sortDirection);
 
     ByteArrayInputStream inputStream = new ByteArrayInputStream(csvData.getBytes(StandardCharsets.UTF_8));
     InputStreamResource resource = new InputStreamResource(inputStream);
@@ -110,25 +114,5 @@ public class IndexDataController implements IndexDataApi {
         .headers(headers)
         .contentType(MediaType.parseMediaType("text/csv"))
         .body(resource);
-  }
-
-  private String convertToCsv(List<IndexData> indexDataList) {
-    String header = "IndexInfoId, BaseDate, ClosingPrice, HighPrice, LowPrice, Variation, FluctuationRate, TradingQuantity, TradingPrice, MarketCapitalization";
-
-    String body = indexDataList.stream()
-        .map(data -> String.join(",",
-            String.valueOf(data.getIndexInfo().getId()),
-            data.getBaseDate().toString(),
-            data.getClosingPrice().toPlainString(),
-            data.getHighPrice().toPlainString(),
-            data.getLowPrice().toPlainString(),
-            data.getVersus().toPlainString(),
-            data.getFluctuationRate().toPlainString(),
-            String.valueOf(data.getTradingQuantity()),
-            String.valueOf(data.getTradingPrice())
-        ))
-        .collect(Collectors.joining("\n"));
-
-    return header + "\n" + body;
   }
 }
